requirements.txt
# Core ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Deep learning
tensorflow>=2.10.0
torch>=1.12.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# API and web
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.10.0

# Utilities
joblib>=1.1.0
pyyaml>=6.0
python-dotenv>=0.20.0
click>=8.1.0

# Monitoring
mlflow>=1.28.0
wandb>=0.13.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
